pyproject.toml
[tool.poetry]
name = "repl-nix-autoloopmediaplayer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
email-validator = "^2.2.0"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
python-magic = "^0.4.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
